Make WordPress artifact helpers self-contained
