+++ wordpress/homeboy.json
@@ -13,6 +13,7 @@
       "node tests/codebox-client-boundary.test.js",
       "node tests/wp-codebox-fuzz-run-smoke.js",
       "node tests/wordpress-fuzz-runner-smoke.js",
+      "node tests/wp-codebox-artifacts-isolated-snapshot-smoke.js",
       "node tests/wordpress-manifest-settings-smoke.js"
     ]
   }
