+++ wordpress/package.json
@@ -105,6 +105,7 @@
     "test:wordpress-fuzz-coverage-aggregate": "node tests/wordpress-fuzz-coverage-aggregate-smoke.js",
     "test:external-http-guardrail": "node tests/external-http-guardrail-smoke.js",
     "test:env-provider": "node tests/env-provider-smoke.js",
+    "test:wp-codebox-artifacts-isolated-snapshot": "node tests/wp-codebox-artifacts-isolated-snapshot-smoke.js",
     "test:wordpress-helper-consumer": "node tests/wordpress-helper-consumer-smoke.mjs",
     "test:materialized-site-quality": "node tests/materialized-site-quality-smoke.js",
     "test:timing-correlator": "node tests/timing-correlator-smoke.js",
